Add a 2,3-b-tree (two three bee minus tree).
This is a fixed order balanced binary tree with favorable logarithmic
complexities, providing a key-value store interface. This structure is
also useful because it is contiguous and uses relative addressing. In
other words: you can create an mmap() backed by a file, then initialize
this structure on it, and you have a persistent key-value database.